Refactor code; create src/components directory
